Refactoring (#177)
* Refactoring

+++ src/Infrastructure/DataAccess/Repositories/AccountRepository.cs
@@ -1,3 +1,7 @@
+// <copyright file="AccountRepository.cs" company="Ivan Paulovich">
+// Copyright © Ivan Paulovich. All rights reserved.
+// </copyright>
+
 namespace Infrastructure.DataAccess.Repositories
 {
     using System;
@@ -15,16 +19,19 @@ namespace Infrastructure.DataAccess.Repositories
     using Credit = Entities.Credit;
     using Debit = Entities.Debit;
 
+    /// <inheritdoc />
     public sealed class AccountRepository : IAccountRepository
     {
         private readonly MangaContext _context;
 
-        public AccountRepository(MangaContext context)
-        {
-            this._context = context ??
-                            throw new ArgumentNullException(nameof(context));
-        }
+        /// <summary>
+        /// </summary>
+        /// <param name="context"></param>
+        public AccountRepository(MangaContext context) => this._context = context ??
+                                                                          throw new ArgumentNullException(
+                                                                              nameof(context));
 
+        /// <inheritdoc />
         public async Task<IList<IAccount>> GetBy(CustomerId customerId)
         {
             var accounts = this._context
@@ -37,6 +44,7 @@ namespace Infrastructure.DataAccess.Repositories
                 .ConfigureAwait(false);
         }
 
+        /// <inheritdoc />
         public async Task Add(IAccount account, ICredit credit)
         {
             await this._context
@@ -50,8 +58,14 @@ namespace Infrastructure.DataAccess.Repositories
                 .ConfigureAwait(false);
         }
 
+        /// <inheritdoc />
         public async Task Delete(IAccount account)
         {
+            if (account is null)
+            {
+                throw new ArgumentNullException(nameof(account));
+            }
+
             const string deleteSQL = @"DELETE FROM Credit WHERE AccountId = @Id;
                       DELETE FROM Debit WHERE AccountId = @Id;
                       DELETE FROM Account WHERE Id = @Id;";
@@ -65,9 +79,10 @@ namespace Infrastructure.DataAccess.Repositories
                 .ConfigureAwait(false);
         }
 
+        /// <inheritdoc />
         public async Task<IAccount> GetAccount(AccountId id)
         {
-            var account = await this._context
+            Account account = await this._context
                 .Accounts
                 .Where(a => a.Id.Equals(id))
                 .SingleOrDefaultAsync()
@@ -75,7 +90,7 @@ namespace Infrastructure.DataAccess.Repositories
 
             if (account is null)
             {
-                return null;
+                return null!;
             }
 
             var credits = this._context
@@ -96,18 +111,14 @@ namespace Infrastructure.DataAccess.Repositories
             return account;
         }
 
-        public async Task Update(IAccount account, ICredit credit)
-        {
-            await this._context
-                .Credits
-                .AddAsync((Credit)credit);
-        }
+        /// <inheritdoc />
+        public async Task Update(IAccount account, ICredit credit) => await this._context
+            .Credits
+            .AddAsync((Credit)credit);
 
-        public async Task Update(IAccount account, IDebit debit)
-        {
-            await this._context
-                .Debits
-                .AddAsync((Debit)debit);
-        }
+        /// <inheritdoc />
+        public async Task Update(IAccount account, IDebit debit) => await this._context
+            .Debits
+            .AddAsync((Debit)debit);
     }
 }
